Split files by platform
This is keeping the same code/functionality, but will make it easier to maintain by splitting the code by platform, so we can focus on the changes on a specific platform without having to change all the code.
This is a preparation work to update the ADC libraries for ESP32, as ESP-IDF v5.1.5 is already informing about a deprecated library in use, but that should be taken in another PR.
