+++ esphome/components/adc/adc_sensor.h
@@ -3,13 +3,12 @@
 #include "esphome/components/sensor/sensor.h"
 #include "esphome/components/voltage_sampler/voltage_sampler.h"
 #include "esphome/core/component.h"
-#include "esphome/core/defines.h"
 #include "esphome/core/hal.h"
 
 #ifdef USE_ESP32
 #include <esp_adc_cal.h>
 #include "driver/adc.h"
-#endif
+#endif  // USE_ESP32
 
 namespace esphome {
 namespace adc {
@@ -43,7 +42,7 @@ class ADCSensor : public sensor::Sensor, public PollingComponent, public voltage
     this->channel1_ = ADC1_CHANNEL_MAX;
   }
   void set_autorange(bool autorange) { this->autorange_ = autorange; }
-#endif
+#endif  // USE_ESP32
 
   /// Update ADC values
   void update() override;
@@ -59,11 +58,11 @@ class ADCSensor : public sensor::Sensor, public PollingComponent, public voltage
 
 #ifdef USE_ESP8266
   std::string unique_id() override;
-#endif
+#endif  // USE_ESP8266
 
 #ifdef USE_RP2040
   void set_is_temperature() { this->is_temperature_ = true; }
-#endif
+#endif  // USE_RP2040
 
  protected:
   InternalGPIOPin *pin_;
@@ -72,7 +71,7 @@ class ADCSensor : public sensor::Sensor, public PollingComponent, public voltage
 
 #ifdef USE_RP2040
   bool is_temperature_{false};
-#endif
+#endif  // USE_RP2040
 
 #ifdef USE_ESP32
   adc_atten_t attenuation_{ADC_ATTEN_DB_0};
@@ -83,8 +82,8 @@ class ADCSensor : public sensor::Sensor, public PollingComponent, public voltage
   esp_adc_cal_characteristics_t cal_characteristics_[SOC_ADC_ATTEN_NUM] = {};
 #else
   esp_adc_cal_characteristics_t cal_characteristics_[ADC_ATTEN_MAX] = {};
-#endif
-#endif
+#endif  // ESP_IDF_VERSION_MAJOR
+#endif  // USE_ESP32
 };
 
 }  // namespace adc
